refactor: split sale landing server/client and harden rent-vs-buy pricing
